+++ CHANGELOG.md
@@ -12,8 +12,12 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 - support for dataset parameters `:using_graph` and `:using_named_graph`
   on update operations
 
-[Compare v0.4.2...HEAD](https://github.com/rdf-elixir/sparql_client/compare/v0.4.2...HEAD)
+### Changed
 
+- graph names on dataset parameters can be provided as RDF namespace terms
+
+
+[Compare v0.4.2...HEAD](https://github.com/rdf-elixir/sparql_client/compare/v0.4.2...HEAD)
 
 
 ## v0.4.2 - 2023-04-24